Adding ASAN coverage to pquery-prep-red.sh, pquery-clean-known++.sh and improving pquery-run.sh

+++ pquery-clean-known++.sh
@@ -23,10 +23,13 @@ ${SCRIPT_PWD}/pquery-results.sh | grep -A1 "Likely 'Server has gone away' 200x d
 #   sed 's|reducers ||;s|,|\n|g' | xargs -I{} ${SCRIPT_PWD}/pquery-del-trial.sh {}
 
 # Delete all 'Assert: no core file found in' trials (benefit of new_text_string.sh)
-${SCRIPT_PWD}/pquery-results.sh | grep "Assert. no core file found in" | grep -o "reducers.*[^)]" | \
- sed 's|reducers ||;s|,|\n|g' | xargs -I{} ${SCRIPT_PWD}/pquery-del-trial.sh {}
+if [ $(grep -m1 --binary-files=text "ERROR:" */log/master.err 2> /dev/null | wc -l) -eq 0 ]; then
+  sleep 0  # Dummy instruction to enable if to remain active irrespective of disabled next line
+  # ${SCRIPT_PWD}/pquery-results.sh | grep "Assert. no core file found in" | grep -o "reducers.*[^)]" | sed 's|reducers ||;s|,|\n|g' | xargs -I{} ${SCRIPT_PWD}/pquery-del-trial.sh {}
+fi
 # 9/9/2020 temp disabled to check why so many 'Assert: no core file found in */*core*' trials
 # 22/9/2020 temp re-enabled (temp re-disable again later) due to so many bugs in queues
+# 11/11/2020 re-disabled
 
 # Delete all 'TRIALS TO CHECK MANUALLY' trials which do not have an associated core file in their data directories
 rm -f ~/results_list++.tmp
